FEAT: 공통 응답 처리 기능 추가

+++ src/main/java/com/team2/levelog/comment/service/CommentService.java
@@ -3,6 +3,8 @@ package com.team2.levelog.comment.service;
 import com.team2.levelog.comment.dto.CommentRequestDto;
 import com.team2.levelog.comment.entity.Comment;
 import com.team2.levelog.comment.repository.CommentRepository;
+import com.team2.levelog.global.GlobalResponse.CustomException;
+import com.team2.levelog.global.GlobalResponse.code.ErrorCode;
 import com.team2.levelog.post.entity.Post;
 import com.team2.levelog.post.repository.PostRepository;
 import com.team2.levelog.user.entity.User;
@@ -11,7 +13,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.List;
+import static com.team2.levelog.global.GlobalResponse.code.ErrorCode.NO_ACCESS;
 
 @Service
 @RequiredArgsConstructor
@@ -23,7 +25,7 @@ public class CommentService {
     // 댓글 작성
     public void createComment(Long postId, CommentRequestDto commentRequestDto, User user){
         Post post = postRepository.findById(postId).orElseThrow(
-                ()-> new IllegalArgumentException("해당 글이 없습니다")
+                ()-> new CustomException(ErrorCode.POST_NOT_FOUND)
         );
 
         Comment newComment = new Comment(post, commentRequestDto, user);
@@ -33,11 +35,11 @@ public class CommentService {
     // 대댓글 작성
     public void createReply(Long postId, Long commentsId, CommentRequestDto commentRequestDto, User user){
         Post post = postRepository.findById(postId).orElseThrow(
-                ()-> new IllegalArgumentException("해당 글이 없습니다")
+                ()-> new CustomException(ErrorCode.POST_NOT_FOUND)
         );
 
         Comment comment = commentRepository.findById(commentsId).orElseThrow(
-                ()-> new IllegalArgumentException("해당 댓글이 없습니다")
+                ()-> new CustomException(ErrorCode.COMMENT_NOT_FOUND)
         );
 
         Comment newComment = new Comment(post, comment, commentRequestDto, user);
@@ -48,8 +50,12 @@ public class CommentService {
     @Transactional
     public void modifyComment(Long commentsId, CommentRequestDto commentRequestDto, User user){
         Comment comment = commentRepository.findById(commentsId).orElseThrow(
-                ()-> new IllegalArgumentException("해당 댓글이 없습니다")
+                ()-> new CustomException(ErrorCode.COMMENT_NOT_FOUND)
         );
+
+        if(!user.getNickname().equals(comment.getNickname())){
+            throw new CustomException(ErrorCode.NO_ACCESS);
+        }
 
         comment.update(commentRequestDto);
     }
@@ -57,8 +63,12 @@ public class CommentService {
     // 댓글, 대댓글 삭제
     public void deleteComment(Long commentsId, User user){
         Comment comment = commentRepository.findById(commentsId).orElseThrow(
-                ()-> new IllegalArgumentException("해당 댓글이 없습니다")
+                ()-> new CustomException(ErrorCode.COMMENT_NOT_FOUND)
         );
+
+        if(!user.getNickname().equals(comment.getNickname())){
+            throw new CustomException(ErrorCode.NO_ACCESS);
+        }
 
         commentRepository.delete(comment);
     }
